+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # Changelog
 
-## Version 13.2.3 - 2024-10-09
+## Version 13.3.0 - 2024-12-03
 
 - Bugfix `transitions update-execution` now supports additional keyword arguments
+- Added optional parameter `--async` to `predictions create`
+- Added `predictions get`
+- Added optional parameter `--statistics-last-n-days` to `models get-training`
+- Added optional parameter `--max-elements` to `models get-training`
+- Remove optional parameter `--in-schema` and `--out-schema` from `transitions create` and `transitions update`
 
 ## Version 13.2.2 - 2024-06-13